pyproject.toml
[tool.poetry]
name = "autonodelabel"
version = "0.1.0"
description = "Automatically label Kubernetes nodes based on their hardware"
authors = ["Jonathan Gazeley <me@jonathangazeley.com>"]

[tool.poetry.dependencies]
python = "^3.11"
py-cpuinfo = "^9.0.0"
kubernetes = "^26.1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
autonodelabel = 'autonodelabel.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
